backend/requirements.txt
# Backend Dependencies for English Teaching Assignment Grading System

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI Integration
litellm==1.23.0
openai==1.12.0
anthropic==0.18.0

# OCR
easyocr==1.7.1
pdf2image==1.16.3
Pillow==10.2.0

# Document Processing
python-docx==1.1.0
reportlab==4.0.9
PyPDF2==3.0.1

# Web Search
duckduckgo-search==4.4.0
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.26.0

# Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0

# Configuration
pyyaml==6.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
uuid6==2024.1.12
